// src/api/apiTypes.tsx
export interface CommentUser {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  profilePicture: string;
}

export interface Comment {
  id: string;
  content: string;
  newsId: string;
  isApproved: boolean;
  isDeleted: boolean;
  createdAt: string;
  updatedAt: string;
  user: CommentUser;
}

export interface CommentsResponse {
  comments: Comment[];
  total: number;
  page: number;
  perPage: number;
  totalPages: number;
}

export interface PostCommentRequest {
  content: string;
  newsId?: string;
}